fix: optimize Cycle size to improve tracer memory usage (#852)
* fix: add a stack canary to prevent stack overflows

* fixes writes to bss

* fix: tracer tests

* fix: wasm

* fix: test

* fix: optimize Cycle size to improve tracer memory usage

* fix fmt

* fix: verifier test fixtures

---------

+++ tracer/src/instruction/lwu.rs
@@ -17,10 +17,10 @@ impl LWU {
     fn exec(&self, cpu: &mut Cpu, ram_access: &mut <LWU as RISCVInstruction>::RAMAccess) {
         // The LWU instruction, on the other hand, zero-extends the 32-bit value from memory for
         // RV64I.
-        let address = cpu.x[self.operands.rs1].wrapping_add(self.operands.imm) as u64;
+        let address = cpu.x[self.operands.rs1 as usize].wrapping_add(self.operands.imm) as u64;
         let value = cpu.mmu.load_word(address);
 
-        cpu.x[self.operands.rd] = match value {
+        cpu.x[self.operands.rd as usize] = match value {
             Ok((word, memory_read)) => {
                 *ram_access = memory_read;
                 // Zero extension for unsigned word load
