+++ tracer/src/instruction/virtual_shift_right_bitmask.rs
@@ -27,14 +27,14 @@ impl VirtualShiftRightBitmask {
     ) {
         match cpu.xlen {
             Xlen::Bit32 => {
-                let shift = cpu.x[self.operands.rs1] as u64 % 32;
+                let shift = cpu.x[self.operands.rs1 as usize] as u64 % 32;
                 let ones = (1u64 << (32 - shift)) - 1;
-                cpu.x[self.operands.rd] = (ones << shift) as i64;
+                cpu.x[self.operands.rd as usize] = (ones << shift) as i64;
             }
             Xlen::Bit64 => {
-                let shift = cpu.x[self.operands.rs1] as u64 % 64;
+                let shift = cpu.x[self.operands.rs1 as usize] as u64 % 64;
                 let ones = (1u128 << (64 - shift)) - 1;
-                cpu.x[self.operands.rd] = (ones << shift) as i64;
+                cpu.x[self.operands.rd as usize] = (ones << shift) as i64;
             }
         }
     }
