+++ tracer/src/instruction/virtual_srai.rs
@@ -19,7 +19,8 @@ declare_riscv_instr!(
 impl VirtualSRAI {
     fn exec(&self, cpu: &mut Cpu, _: &mut <VirtualSRAI as RISCVInstruction>::RAMAccess) {
         let shift = self.operands.imm.trailing_zeros();
-        cpu.x[self.operands.rd] = cpu.sign_extend(cpu.x[self.operands.rs1].wrapping_shr(shift));
+        cpu.x[self.operands.rd as usize] =
+            cpu.sign_extend(cpu.x[self.operands.rs1 as usize].wrapping_shr(shift));
     }
 }
